[#549] Move layer logic to handler
