+++ api/handler/detector.go
@@ -1,4 +1,4 @@
-package layer
+package handler
 
 import (
 	"io"
