+++ api/handler/neofs.go
@@ -1,4 +1,4 @@
-package layer
+package handler
 
 import (
 	"context"
