+++ api/handler/neofs_mock.go
@@ -1,4 +1,4 @@
-package layer
+package handler
 
 import (
 	"bytes"
