+++ api/handler/util.go
@@ -2,15 +2,21 @@ package handler
 
 import (
 	"context"
+	"encoding/hex"
 	errorsStd "errors"
+	"fmt"
 	"net/http"
+	"os"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/nspcc-dev/neofs-s3-gw/api"
 	"github.com/nspcc-dev/neofs-s3-gw/api/data"
 	"github.com/nspcc-dev/neofs-s3-gw/api/errors"
-	"github.com/nspcc-dev/neofs-s3-gw/api/layer"
+	"github.com/nspcc-dev/neofs-s3-gw/api/handler/encryption"
+	"github.com/nspcc-dev/neofs-s3-gw/creds/accessbox"
+	"github.com/nspcc-dev/neofs-sdk-go/object"
 	"github.com/nspcc-dev/neofs-sdk-go/session"
 	"go.uber.org/zap"
 )
@@ -34,8 +40,8 @@ func transformToS3Error(err error) error {
 		return err
 	}
 
-	if errorsStd.Is(err, layer.ErrAccessDenied) ||
-		errorsStd.Is(err, layer.ErrNodeAccessDenied) {
+	if errorsStd.Is(err, ErrAccessDenied) ||
+		errorsStd.Is(err, ErrNodeAccessDenied) {
 		return errors.GetAPIError(errors.ErrAccessDenied)
 	}
 
@@ -43,7 +49,7 @@ func transformToS3Error(err error) error {
 }
 
 func (h *handler) getBucketAndCheckOwner(r *http.Request, bucket string, header ...string) (*data.BucketInfo, error) {
-	bktInfo, err := h.obj.GetBucketInfo(r.Context(), bucket)
+	bktInfo, err := h.getBucketInfo(r.Context(), bucket)
 	if err != nil {
 		return nil, err
 	}
@@ -62,7 +68,7 @@ func (h *handler) getBucketAndCheckOwner(r *http.Request, bucket string, header 
 	return bktInfo, checkOwner(bktInfo, expected)
 }
 
-func parseRange(s string) (*layer.RangeParams, error) {
+func parseRange(s string) (*RangeParams, error) {
 	if s == "" {
 		return nil, nil
 	}
@@ -92,14 +98,14 @@ func parseRange(s string) (*layer.RangeParams, error) {
 		return nil, errors.GetAPIError(errors.ErrInvalidRange)
 	}
 
-	return &layer.RangeParams{
+	return &RangeParams{
 		Start: values[0],
 		End:   values[1],
 	}, nil
 }
 
 func getSessionTokenSetEACL(ctx context.Context) (*session.Container, error) {
-	boxData, err := layer.GetBoxData(ctx)
+	boxData, err := GetBoxData(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -109,4 +115,140 @@ func getSessionTokenSetEACL(ctx context.Context) (*session.Container, error) {
 	}
 
 	return sessionToken, nil
+}
+
+type (
+	// ListObjectsInfo contains common fields of data for ListObjectsV1 and ListObjectsV2.
+	ListObjectsInfo struct {
+		Prefixes    []string
+		Objects     []*data.ObjectInfo
+		IsTruncated bool
+	}
+
+	// ListObjectsInfoV1 holds data which ListObjectsV1 returns.
+	ListObjectsInfoV1 struct {
+		ListObjectsInfo
+		NextMarker string
+	}
+
+	// ListObjectsInfoV2 holds data which ListObjectsV2 returns.
+	ListObjectsInfoV2 struct {
+		ListObjectsInfo
+		NextContinuationToken string
+	}
+
+	// ListObjectVersionsInfo stores info and list of objects versions.
+	ListObjectVersionsInfo struct {
+		CommonPrefixes      []string
+		IsTruncated         bool
+		KeyMarker           string
+		NextKeyMarker       string
+		NextVersionIDMarker string
+		Version             []*data.ExtendedObjectInfo
+		DeleteMarker        []*data.ExtendedObjectInfo
+		VersionIDMarker     string
+	}
+)
+
+// PathSeparator is a path components separator string.
+const PathSeparator = string(os.PathSeparator)
+
+func userHeaders(attrs []object.Attribute) map[string]string {
+	result := make(map[string]string, len(attrs))
+
+	for _, attr := range attrs {
+		result[attr.Key()] = attr.Value()
+	}
+
+	return result
+}
+
+func objectInfoFromMeta(bkt *data.BucketInfo, meta *object.Object) *data.ObjectInfo {
+	var (
+		mimeType string
+		creation time.Time
+	)
+
+	headers := userHeaders(meta.Attributes())
+	delete(headers, object.AttributeFilePath)
+	if contentType, ok := headers[object.AttributeContentType]; ok {
+		mimeType = contentType
+		delete(headers, object.AttributeContentType)
+	}
+	if val, ok := headers[object.AttributeTimestamp]; !ok {
+		// ignore empty value
+	} else if dt, err := strconv.ParseInt(val, 10, 64); err == nil {
+		creation = time.Unix(dt, 0)
+		delete(headers, object.AttributeTimestamp)
+	}
+
+	objID, _ := meta.ID()
+	payloadChecksum, _ := meta.PayloadChecksum()
+	return &data.ObjectInfo{
+		ID:    objID,
+		CID:   bkt.CID,
+		IsDir: false,
+
+		Bucket:      bkt.Name,
+		Name:        filepathFromObject(meta),
+		Created:     creation,
+		ContentType: mimeType,
+		Headers:     headers,
+		Owner:       *meta.OwnerID(),
+		Size:        int64(meta.PayloadSize()),
+		HashSum:     hex.EncodeToString(payloadChecksum.Value()),
+	}
+}
+
+func FormEncryptionInfo(headers map[string]string) encryption.ObjectEncryption {
+	algorithm := headers[AttributeEncryptionAlgorithm]
+	return encryption.ObjectEncryption{
+		Enabled:   len(algorithm) > 0,
+		Algorithm: algorithm,
+		HMACKey:   headers[AttributeHMACKey],
+		HMACSalt:  headers[AttributeHMACSalt],
+	}
+}
+
+func addEncryptionHeaders(meta map[string]string, enc encryption.Params) error {
+	meta[AttributeEncryptionAlgorithm] = AESEncryptionAlgorithm
+	hmacKey, hmacSalt, err := enc.HMAC()
+	if err != nil {
+		return fmt.Errorf("get hmac: %w", err)
+	}
+	meta[AttributeHMACKey] = hex.EncodeToString(hmacKey)
+	meta[AttributeHMACSalt] = hex.EncodeToString(hmacSalt)
+
+	return nil
+}
+
+func filepathFromObject(o *object.Object) string {
+	for _, attr := range o.Attributes() {
+		if attr.Key() == object.AttributeFilePath {
+			return attr.Value()
+		}
+	}
+	objID, _ := o.ID()
+	return objID.EncodeToString()
+}
+
+// NameFromString splits name into a base file name and a directory path.
+func NameFromString(name string) (string, string) {
+	ind := strings.LastIndex(name, PathSeparator)
+	return name[ind+1:], name[:ind+1]
+}
+
+// GetBoxData  extracts accessbox.Box from context.
+func GetBoxData(ctx context.Context) (*accessbox.Box, error) {
+	var boxData *accessbox.Box
+	data, ok := ctx.Value(api.BoxData).(*accessbox.Box)
+	if !ok || data == nil {
+		return nil, fmt.Errorf("couldn't get box data from context")
+	}
+
+	boxData = data
+	if boxData.Gate == nil {
+		boxData.Gate = &accessbox.GateData{}
+	}
+	return boxData, nil
 }
