+++ internal/neofs/neofs_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/nspcc-dev/neofs-s3-gw/api/layer"
+	"github.com/nspcc-dev/neofs-s3-gw/api/handler"
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
 	"github.com/stretchr/testify/require"
 )
@@ -17,9 +17,9 @@ func TestErrorChecking(t *testing.T) {
 	var wrappedError error
 
 	if fetchedReason, ok := isErrAccessDenied(err); ok {
-		wrappedError = fmt.Errorf("%w: %s", layer.ErrAccessDenied, fetchedReason)
+		wrappedError = fmt.Errorf("%w: %s", handler.ErrAccessDenied, fetchedReason)
 	}
 
-	require.ErrorIs(t, wrappedError, layer.ErrAccessDenied)
+	require.ErrorIs(t, wrappedError, handler.ErrAccessDenied)
 	require.Contains(t, wrappedError.Error(), reason)
 }
